requirements.txt
asgiref==3.5.0
astroid==2.9.3
black==22.1.0
cffi==1.15.0
cfgv==3.3.1
click==8.0.3
cryptography==36.0.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
gunicorn==20.1.0
identify==2.4.7
importlib-metadata==4.10.1
isort==5.10.1
lazy-object-proxy==1.7.1
libcst==0.3.23
Markdown==3.3.6
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.4.1
pre-commit==2.17.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycln==1.1.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
typer==0.4.0
typing-inspect==0.7.1
typing_extensions==4.0.1
virtualenv==20.13.0
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.7.0
